redis is a viable transport for local dev, not production envs.
Basically redis as a transport is not secure, only used by a handful of people in the community run it, does not scale as well, and fundementally is wrong solution to what the transport should do. The only reason it was considered was for reducing the complexity of the setup for **local development**.

We should encourage people more strongly **especially** new members to avoid this as it is a terrible idea.

Signed-off-by: Ben Abrams <me@benabrams.it>